tests/wpt/tests/speculation-rules/prerender/restrictions_shared_storage_worklet.https.html

<!DOCTYPE html>
<title>Worklet operations to Shared Storage API is deferred</title>
<meta name="timeout" content="long">
<meta name="variant" content="?type=add_module">
<meta name="variant" content="?type=select_url">
<meta name="variant" content="?type=run">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());
promise_test(async t => {
    const uid = token();
    const bc = new PrerenderChannel('test-channel', uid);

    const gotMessage = new Promise(resolve => {
      bc.addEventListener('message', e => {
        resolve(e.data);
      }, {
        once: true
      });
    });

    const params = new URLSearchParams(location.search);
    const test_type = params.get('type');
    const url =
      `resources/shared-storage.https.html?uid=${uid}&type=${test_type}`;
    window.open(url, '_blank', 'noopener');

    const result = await gotMessage;
    const expected = [{
        event: `started waiting ${test_type}`,
        prerendering: true
      },
      {
        event: 'prerendering change',
        prerendering: false
      },
      {
        event: `finished waiting ${test_type}`,
        prerendering: false
      }
    ];
    assert_equals(result.length, expected.length);
    for (let i = 0; i < result.length; i++) {
      assert_equals(result[i].event, expected[i].event,
        `event[${i}]`);
      assert_equals(result[i].prerendering, expected[i].prerendering,
        `prerendering[${i}]`);
    }
    bc.close();

    // Send a close signal to PrerenderEventCollector on the prerendered page.
    new PrerenderChannel('close', uid).postMessage('');
  },
  `Worklet operations the Shared Storage API should
   be deferred until the prerendered page is activated`);
</script>
